+++ src/main/scala/com/atomist/rug/runtime/js/JavaScriptInvokingProjectOperation.scala
@@ -181,7 +181,13 @@ abstract class JavaScriptInvokingProjectOperation(
     parameter.setDefaultRef(details.get("defaultRef").asInstanceOf[String])
     val disp = details.get("displayable")
     parameter.setDisplayable(if(disp != null) disp.asInstanceOf[Boolean] else true)
-    parameter.setRequired(details.get("required").asInstanceOf[Boolean])
+
+    if(details.hasMember("required")){
+      parameter.setRequired(details.get("required").asInstanceOf[Boolean])
+    }else{
+      parameter.setRequired(true)
+    }
+
 
     parameter.addTags(readTagsFromMetadata(details))